+++ loki/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.3
+
+* Removed Grafana to separate flavour
+
+---
+
 0.0.2
 
 * Addition of Grafana Loki